LayoutTests/fast/canvas/webgl/gl-getshadersource.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"
  "http://www.w3.org/TR/html4/loose.dtd">
<html>
<head>
<meta content='text/html; charset=UTF-8' http-equiv='Content-Type'/>
   <title>WebGL getShaderSource conformance test.</title>
<link rel="stylesheet" href="../../js/resources/js-test-style.css"/>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/webgl-test.js"> </script>
<script src="resources/webgl-test-utils.js"> </script>
</head>
<body>
<canvas id="example" width="4" height="4" style="width: 40px; height: 30px;"></canvas>
<div id="description"></div>
<div id="console"></div>
<script id="vshader" type="x-shader/x-vertex">abc//defＮＯＴＡＳＣＩＩ</script>
<script>
description("Tests that the source that goes into a shader is what comes out.");
var wtu = WebGLTestUtils;
var canvas = document.getElementById("example");
var gl = wtu.create3DContext(canvas);
var original = document.getElementById("vshader").text;
var shader = gl.createShader(gl.VERTEX_SHADER);
var sourceLength = gl.getShaderParameter(shader, gl.SHADER_SOURCE_LENGTH);
shouldBe("sourceLength", "0");
gl.shaderSource(shader, original);
var source = gl.getShaderSource(shader);
shouldBe("source", "original");
sourceLength = gl.getShaderParameter(shader, gl.SHADER_SOURCE_LENGTH);
shouldBe("sourceLength", "17");
successfullyParsed = true;
glErrorShouldBe(gl, gl.NO_ERROR, "Should be no errors.");
</script>
</body>
</body>
<script src="../../js/resources/js-test-post.js"></script>

<script>
</script>

</body>
</html>
